[fix] #176 AI 블록 추천 관련 문제 해결
